backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
python-dotenv==1.0.1
google-genai==1.0.0
supabase==2.10.0
pandas==3.0.1
python-multipart==0.0.9
